requirements.txt
# Web scraping and RSS feeds
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
PyYAML>=6.0

# Environment / secrets
python-dotenv>=1.0.0

# API clients
openai>=1.0.0  # Required for auto-ingest claim extraction
# anthropic>=0.18.0  # Optional - for alternative claim extraction

# Existing optional dependencies
matplotlib>=3.7.0
seaborn>=0.12.0
